Fix return type by the method in the game difficulty tier script

+++ Assets/Game/Scripts/GameDifficultyTier.cs
@@ -7,6 +7,6 @@ public class GameDifficultyTier
 	[SerializeField, Min(1)] private int enemiesLimitAtOnce;
 	[SerializeField, Min(0.01f)] private float enemyMovementSpeedMultiplier;
 
-	public float GetEnemiesLimitAtOnce() => enemiesLimitAtOnce;
+	public int GetEnemiesLimitAtOnce() => enemiesLimitAtOnce;
 	public float GetEnemyMovementSpeedMultiplier() => enemyMovementSpeedMultiplier;
 }
